requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
paho-mqtt==1.5.1
Werkzeug==2.0.1
psutil==5.9.0
python-dotenv==0.19.0
gunicorn==20.1.0
eventlet==0.30.2